add reinitialization to the form in modal and move helpers to the separate file
