+++ packages/frinx-workflow-ui/src/utils/helpers.utils.ts
@@ -1,5 +1,7 @@
 import { Workflow } from '../helpers/types';
 
+export type InputParameter = Record<string, { value: string; description: string; type: string }>;
+
 export const sortAscBy = (key: string) => {
   return function (x: Record<string, any>, y: Record<string, any>) {
     return x[key] === y[key] ? 0 : x[key] > y[key] ? 1 : -1;
@@ -31,3 +33,51 @@ export const getDynamicInputParametersFromWorkflow = (workflow?: Workflow | null
 
   return match || [];
 };
+
+export function parseInputParameters(inputParameters?: string[]) {
+  if (inputParameters == null || inputParameters.length === 0) {
+    return null;
+  }
+
+  const parsedInputParameters: InputParameter[] = inputParameters.map(jsonParse);
+
+  return parsedInputParameters.map(Object.keys).reduce((acc, currObjectKeys, index) => {
+    const result: InputParameter = currObjectKeys.reduce(
+      (acc, curr) => ({ ...acc, [curr]: parsedInputParameters[index][curr] }),
+      {},
+    );
+
+    return {
+      ...acc,
+      ...result,
+    };
+  }, {} as InputParameter);
+}
+
+export function getInitialValuesFromParsedInputParameters(
+  parsedInputParameters?: InputParameter | null,
+  dynamicInputParameters?: string[] | null,
+) {
+  let initialValues = {};
+  if (parsedInputParameters != null) {
+    initialValues = Object.keys(parsedInputParameters).reduce(
+      (acc, curr) => ({ ...acc, [curr]: parsedInputParameters[curr].value }),
+      {},
+    );
+  }
+
+  if (dynamicInputParameters != null) {
+    initialValues = {
+      ...dynamicInputParameters?.reduce(
+        (acc, curr) => ({
+          ...acc,
+          [curr]: '',
+        }),
+        {},
+      ),
+      ...initialValues,
+    };
+  }
+
+  return initialValues;
+}
